фикс даты оплаты заказа, фикс массы заказа, валидация даты доставки, null safety createOrderFromCart

+++ domain/src/main/java/com/lakomka/dto/OrderXmlDto.java
@@ -30,7 +30,7 @@ public class OrderXmlDto {
     String dateTimeOrder;
 
     BigDecimal sumOrder;
-    Integer sumWeight;
+    Double sumWeight;
 
     boolean bitAccPrint;
     boolean bitSertifPrint;
